У частичек своя сетка

+++ src/Vge/World/WorldBase.cs
@@ -439,7 +439,7 @@ namespace Vge.World
         /// Заспавнить частицы
         /// </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public virtual void SpawnParticle(int particleId, 
+        public virtual void SpawnParticle(ushort particleId, 
             int count, Vector3 pos, Vector3 offset, float motion, int parameter) { }
 
         #endregion
