+++ src/Vge/World/WorldClient.cs
@@ -203,7 +203,7 @@ namespace Vge.World
         /// Заспавнить частицы
         /// </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public override void SpawnParticle(int particleId,
+        public override void SpawnParticle(ushort particleId,
             int count, Vector3 pos, Vector3 offset, float motion, int parameter)
         {
             ParticlesRenderer particles = Game.WorldRender.Particles;
